geo::algorithm::ball_tree

Trait BallTreePoint

pub trait BallTreePoint<T: GeoFloat> {
    type Data;

    // Required methods
    fn coord(&self) -> Coord<T>;
    fn into_point_data(self) -> (Point<T>, Self::Data);
}

A point-like value that can be indexed in a BallTree.

Implemented for Point, Coord, and PointWithData. A collection of any BallTreePoint may be passed to BallTree::new or BallTreeBuilder::build. The associated Data type determines the payload returned from queries: it is () for bare points, and D for PointWithData<T, D>.

Required Associated Types

type Data

Associated data carried alongside each point. () for bare points.

Required Methods

fn coord(&self) -> Coord<T>

The 2-D coordinate used for distance calculations during tree construction and queries.

fn into_point_data(self) -> (Point<T>, Self::Data)

Consume self, yielding its point and associated data.