geo::algorithm

Module buffer

Create a new geometry whose boundary is offset the specified distance from the input.

Structs

BufferStyle

Configuration for buffer styling operations.

Enums

LineCap

The endpoint style of a line.

LineJoin

The join style of a line.

Traits

Buffer

Create a new geometry whose boundary is offset the specified distance from the input.