geo::algorithm::contains_properly

Trait ContainsProperly

pub trait ContainsProperly<Rhs = Self> {
    // Required method
    fn contains_properly(&self, rhs: &Rhs) -> bool;
}

Checks if rhs is completely contained within the interior of self.

More formally, the interior of rhs has non-empty (set-theoretic) intersection but neither the interior, nor the boundary of rhs intersects the boundary nor exterior of self. In other words, the DE-9IM intersection matrix of (rhs, self) is T**FF*FF*.

Examples

use geo::ContainsProperly;
use geo::{line_string, point, Polygon};

let line_string = line_string![
    (x: 0., y: 0.),
    (x: 2., y: 0.),
    (x: 2., y: 2.),
    (x: 0., y: 2.),
    (x: 0., y: 0.),
];

let polygon = Polygon::new(line_string.clone(), vec![]);

// Point in Point
assert!(point!(x: 2., y: 0.).contains_properly(&point!(x: 2., y: 0.)));

// Point in Linestring
assert!(line_string.contains_properly(&point!(x: 2., y: 0.)));

// Point in Polygon
assert!(!polygon.contains_properly(&point!(x: 2., y: 0.)));
assert!(polygon.contains_properly(&point!(x: 1., y: 1.)));

Performance Note

Much of this trait is currently implemented by delegating to the Relate trait - see IntersectionMatrix::is_contains_properly; ContainsProperly is faster when checking between Polygon and MultiPolygon when inputs are smaller than about 650 vertices, otherwise use Relate::relate().is_contains_properly.

Required Methods

fn contains_properly(&self, rhs: &Rhs) -> bool