geo::algorithm::covers

Trait Covers

pub trait Covers<Rhs = Self> {
    // Required method
    fn covers(&self, rhs: &Rhs) -> bool;
}

Checks if every point in rhs lies inside self. (i.e. intersects the interior or boundary of) self. Equivalently, tests that no point of other lies outside (in the exterior of) self In other words, the DE-9IM intersection matrix of (rhs, self) matches one of the following patterns: [T*****FF*], [*T****FF*], [***T**FF*], [****T*FF*]

Examples

use geo::Covers;
use geo::{line_string, point, Polygon};

let line_string = line_string![
    (x: 0., y: 0.),
    (x: 2., y: 0.),
    (x: 2., y: 2.),
    (x: 0., y: 2.),
    (x: 0., y: 0.),
];

let polygon = Polygon::new(line_string.clone(), vec![]);

// Point in Point
assert!(point!(x: 2., y: 0.).covers(&point!(x: 2., y: 0.)));

// Point in Linestring
assert!(line_string.covers(&point!(x: 2., y: 0.)));

// Point in Polygon
assert!(polygon.covers(&point!(x: 1., y: 1.)));

Performance Note

Much of this trait is currently implemented by delegating to the Relate trait - see IntersectionMatrix::is_covers; However, Covers may be faster for checking if geometries are covered by Point, Line, Rect and Triangle

Required Methods

fn covers(&self, rhs: &Rhs) -> bool