geo::algorithm::haversine_bearing

Trait HaversineBearing

pub trait HaversineBearing<T: CoordFloat> {
    // Required method
    fn haversine_bearing(&self, point: Point<T>) -> T;
}

👎Deprecated since 0.29.0: Please use the Haversine.bearing method from the Bearing trait instead

Returns the bearing to another Point in degrees.

Bullock, R.: Great Circle Distances and Bearings Between Two Locations, 2007. (https://dtcenter.org/met/users/docs/write_ups/gc_simple.pdf)

Required Methods

fn haversine_bearing(&self, point: Point<T>) -> T

👎Deprecated since 0.29.0: Please use the Haversine.bearing method from the Bearing trait instead

Returns the bearing to another Point in degrees, where North is 0° and East is 90°.

Examples

use geo::HaversineBearing;
use geo::Point;

let p_1 = Point::new(9.177789688110352, 48.776781529534965);
let p_2 = Point::new(9.274410083250379, 48.84033282787534);
let bearing = p_1.haversine_bearing(p_2);
assert_relative_eq!(bearing, 45., epsilon = 1.0e-6);