geo::algorithm::haversine_distance

Trait HaversineDistance

pub trait HaversineDistance<T, Rhs = Self> {
    // Required method
    fn haversine_distance(&self, rhs: &Rhs) -> T;
}

👎Deprecated since 0.29.0: Please use the Haversine.distance method from the Distance trait instead

Determine the distance between two geometries using the haversine formula.

Note: this implementation uses a mean earth radius of 6371.0088 km (6_371_008.7714 m), based on the recommendation of the IUGG:

Moritz, H. (2000). Geodetic Reference System 1980. Journal of Geodesy, 74(1), 128–133. doi:10.1007/s001900050278 “Derived Geometric Constants: R1: mean radius” (p131)

https://link.springer.com/article/10.1007%2Fs001900050278

https://sci-hub.se/https://doi.org/10.1007/s001900050278

Required Methods

fn haversine_distance(&self, rhs: &Rhs) -> T

👎Deprecated since 0.29.0: Please use the Haversine.distance method from the Distance trait instead

Determine the distance between two geometries using the haversine formula.

Units

return value: meters

Examples

use geo::prelude::*;
use geo::point;

// New York City
let p1 = point!(x: -74.006f64, y: 40.7128f64);

// London
let p2 = point!(x: -0.1278f64, y: 51.5074f64);

let distance = p1.haversine_distance(&p2);

assert_eq!(
    5_570_230., // meters
    distance.round()
);