geo::algorithm

Module indexed

Geometries backed by spatial indexes for fast bulk operations

Re-exports

pub use prepared_geometry::PreparedGeometry;

pub use interval_tree_multipolygon::IntervalTreeMultiPolygon;

Modules

interval_tree_multipolygon

A MultiPolygon backed by an interval tree spatial index for fast point containment queries

prepared_geometry

Geometries that are backed by an R*-tree spatial index for faster bulk operations