geo::algorithm::indexed

Module interval_tree_multipolygon

A MultiPolygon backed by an interval tree spatial index for fast point containment queries

Structs

IntervalTreeMultiPolygon

A MultiPolygon backed by an interval tree for fast containment queries