geo::algorithm::indexed::prepared_geometry

Struct PreparedGeometry

pub struct PreparedGeometry<'a, G, F = f64>
where
    G: Into<GeometryCow<'a, F>>,
    F: GeoFloat + RTreeNum,
{ /* private fields */ }

A PreparedGeometry is backed by an R*-tree spatial index and can be more efficient than a plain Geometry when performing multiple topological comparisons against the PreparedGeometry.

use geo::{Relate, PreparedGeometry, wkt};

let polygon = wkt! { POLYGON((2.0 2.0,2.0 6.0,4.0 6.0)) };
let touching_line = wkt! { LINESTRING(0.0 0.0,2.0 2.0) };
let intersecting_line = wkt! { LINESTRING(0.0 0.0,3.0 3.0) };
let contained_line = wkt! { LINESTRING(2.0 2.0,3.0 5.0) };

let prepared_polygon = PreparedGeometry::from(polygon);
assert!(prepared_polygon.relate(&touching_line).is_touches());
assert!(prepared_polygon.relate(&intersecting_line).is_intersects());
assert!(prepared_polygon.relate(&contained_line).is_contains());

Implementations

impl<'a, G, F> PreparedGeometry<'a, G, F>
where F: GeoFloat + RTreeNum, G: Into<GeometryCow<'a, F>>,

pub fn geometry(&self) -> &G

pub fn into_geometry(self) -> G