Trait geo::algorithm::intersects::Intersects

pub trait Intersects<Rhs = Self> {
    fn intersects(&self, rhs: &Rhs) -> bool;
}

Checks if the geometry A intersects the geometry B.

Required Methods

fn intersects(&self, rhs: &Rhs) -> bool

Checks if the geometry A intersects the geometry B.

use geo::{Coordinate, Point, LineString};
use geo::algorithm::intersects::Intersects;

let p = |x, y| Point(Coordinate { x: x, y: y });
let linestring = LineString(vec![p(3., 2.), p(7., 6.)]);

assert!(linestring.intersects(&LineString(vec![p(3., 4.), p(8., 4.)])));
assert!(!linestring.intersects(&LineString(vec![p(9., 2.), p(11., 5.)])));