Struct geo::algorithm::kernels::simple::SimpleKernel

pub struct SimpleKernel;

Simple kernel provides the direct implementation of the predicates. These are meant to be used with exact arithmetic signed types (eg. i32, i64).