geo::algorithm::line_measures

Module metric_spaces

Structs

Euclidean

Operations on the Euclidean plane measure distance with the pythagorean formula - what you’d measure with a ruler.

GeodesicMeasure

Use the Geodesic constant (an instance of GeodesicMeasure) rather than building your own customized GeodesicMeasure for standard ellipsoidal Earth measurements.

HaversineMeasure

Use the Haversine constant (an instance of HaversineMeasure) rather than building your own customized HaversineMeasure for standard spherical Earth measurements.

Rhumb

Provides rhumb line (a.k.a. loxodrome) geometry operations. A rhumb line appears as a straight line on a Mercator projection map.

Constants

Haversine

Haversine measures distance on a sphere using the haversine formula. Distances are considered great circle lengths and given in meters.

Statics

Geodesic

Measures distance on an ellipsoidal model of the Earth using methods given by Karney (2013).