geo::algorithm::line_measures::metric_spaces

Struct GeodesicMeasure

pub struct GeodesicMeasure<F = fn() -> Geodesic> { /* private fields */ }

Use the Geodesic constant (an instance of GeodesicMeasure) rather than building your own customized GeodesicMeasure for standard ellipsoidal Earth measurements.

Measures distance on an ellipsoidal model of the Earth using methods given by Karney (2013).

Geodesic measurements are more accurate than Haversine, but a bit slower, because they account for the fact that the Earth is not a perfect sphere.

Distances are computed using geodesic lines and are measured in meters.

GeodesicMeasure allows creating a geodesic model with custom values for the equatorial radius (A) and the inverse flattening factor (F). Unless you have special needs, use Geodesic, which is simpler than this struct.

Example

use geo::{wkt, Geodesic, GeodesicMeasure, Distance};

let start = wkt!(POINT(23.319941 42.698334)); // Sofia: Longitude, Latitude
let finish = wkt!(POINT(24.742168 42.136097)); // Plovdiv: Longitude, Latitude

// Typically you can use `Geodesic` for Earth measurement, which uses the wgs84 geoid.
assert_relative_eq!(
  132675.5018588206, // meters
  Geodesic.distance(start, finish)
);

// If you have special needs, you can use a custom Earth radius/flattening:
let nad83_equatorial_radius = 6378137.0;
let nad83_flattening = 1. / 298.257222101;
let nad83_geoid = GeodesicMeasure::new(nad83_equatorial_radius, nad83_flattening);
// ever so slightly different from the wgs84 default.
assert_relative_eq!(
  132675.50185928209, // meters
  nad83_geoid.distance(start, finish)
);

// Or you can specify whatever radius you want to get some "out of this world" results. 👽
// from https://nssdc.gsfc.nasa.gov/planetary/factsheet/marsfact.html
let mars_equatorial_radius = 3_396_200.0; // meters
let mars_flattening = 0.00589;
let mars_geoid = GeodesicMeasure::new(mars_equatorial_radius, mars_flattening);
assert_relative_eq!(
  70684.36315529353, // meters
  mars_geoid.distance(start, finish)
);

Implementations

impl GeodesicMeasure

pub const fn wgs84() -> Self

impl GeodesicMeasure<Box<dyn FnOnce() -> Geodesic>>

pub fn new(equatorial_radius: f64, inverse_flattening: f64) -> Self