geo::algorithm::line_measures::metric_spaces

Struct HaversineMeasure

pub struct HaversineMeasure { /* private fields */ }

Use the Haversine constant (an instance of HaversineMeasure) rather than building your own customized HaversineMeasure for standard spherical Earth measurements.

HaversineMeasure measures distance on a sphere using the haversine formula. Distances are considered great circle lengths and given in units that match those of the radius passed to HaversineMeasure::new (typically meters).

You may specify a custom radius for the Earth (or other sphere), but for normal spherical measurements of the Earth, you should use the simpler Haversine which uses the standard earth radius of 6371.0088 km (6_371_008.7714 m), based on the recommendation of the IUGG.

Examples

use geo::{wkt, HaversineMeasure, Haversine, Distance};

let start = wkt!(POINT(23.319941 42.698334)); // Sofia: Longitude, Latitude
let finish = wkt!(POINT(24.742168 42.136097)); // Plovdiv: Longitude, Latitude

// Typically, you can use `Haversine` for measuring on the Earth's surface.
assert_relative_eq!(
    132433.09929460194,
    Haversine.distance(start, finish)
);

// The default `Haversine` const has a radius equal to the mean radius of the GRS80 ellipsoid (6371.0088 km).
assert_relative_eq!(
    Haversine.radius(),
    HaversineMeasure::GRS80_MEAN_RADIUS.radius()
);

// You may choose to use one of the other well known estimations of the Earth's radius,
// which may result in *slightly* different results.
assert_relative_eq!(
    132433.06564071847,
    HaversineMeasure::GRS80_EQUAL_AREA.distance(start, finish)
);

// Or you can specify whatever radius you want to get some "out of this world" results.
let mars_sphere = HaversineMeasure::new(3_389_500.0); // 👽 Mars radius in meters
assert_relative_eq!(
    70456.97222377927,
    mars_sphere.distance(start, finish)
);

References

Moritz, H. (2000). Geodetic Reference System 1980. Journal of Geodesy, 74(1), 128–133. doi:10.1007/s001900050278 “Derived Geometric Constants: R1: mean radius” (p131)

https://link.springer.com/article/10.1007%2Fs001900050278

https://sci-hub.se/https://doi.org/10.1007/s001900050278

Implementations

impl HaversineMeasure

pub const GRS80_MEAN_RADIUS: Self

A sphere with radius equal to the mean radius of the GRS80 ellipsoid — R₁, copied from Moritz (2000).

Moritz, H. (2000). Geodetic Reference System 1980. Journal of Geodesy, 74(1), 128–133. doi:10.1007/s001900050278 “Derived Geometric Constants: R₁: mean radius” (p131)

https://link.springer.com/article/10.1007%2Fs001900050278

https://sci-hub.se/https://doi.org/10.1007/s001900050278

pub const GRS80_EQUAL_AREA: Self

A sphere with the same surface area as the GRS80 ellipsoid, having radius R₂, copied from Moritz (2000).

Moritz, H. (2000). Geodetic Reference System 1980. Journal of Geodesy, 74(1), 128–133. doi:10.1007/s001900050278 “Derived Geometric Constants: R₂: radius of sphere of same surface” (p131)

https://link.springer.com/article/10.1007%2Fs001900050278

https://sci-hub.se/https://doi.org/10.1007/s001900050278

pub const GRS80_EQUAL_VOLUME: Self

A sphere with the same volume as the GRS80 ellipsoid, having radius R₃, copied from Moritz (2000).

Moritz, H. (2000). Geodetic Reference System 1980. Journal of Geodesy, 74(1), 128–133. doi:10.1007/s001900050278 “Derived Geometric Constants: R₃: radius of sphere of same volume” (p131)

https://link.springer.com/article/10.1007%2Fs001900050278

https://sci-hub.se/https://doi.org/10.1007/s001900050278

pub const fn new(radius: f64) -> Self

Parameters

radius: The radius of the sphere, typically in meters.

pub const fn radius(&self) -> f64