geo::algorithm::line_measures::metric_spaces

Struct Rhumb

pub struct Rhumb;

Provides rhumb line (a.k.a. loxodrome) geometry operations. A rhumb line appears as a straight line on a Mercator projection map.

Rhumb distance is measured in meters.

References

The distance, destination, and bearing implementations are adapted in part from their equivalents in Turf.js, which in turn are adapted from the Movable Type spherical geodesy tools.

Turf.js is copyright its authors and the geodesy tools are copyright Chris Veness; both are available under an MIT license.