geo::algorithm::line_measures

Trait Densifiable

pub trait Densifiable<F: CoordFloat> {
    type Output;

    // Required method
    fn densify<MetricSpace>(
        &self,
        metric_space: &MetricSpace,
        max_segment_length: F,
    ) -> Self::Output
       where MetricSpace: Distance<F, Point<F>, Point<F>> + InterpolatePoint<F>;
}

Creates a copy of the geometry with additional points inserted as necessary to ensure there is never more than max_segment_length between points.

It’s typically more convenient to use the Densify trait instead of this trait directly.

Units

max_segment_length units depend on the implementing metric space. It must be greater than 0.

Examples

use geo::wkt;
use geo::line_measures::{Euclidean, Densifiable};

let line_string = wkt!(LINESTRING(0.0 0.0,0.0 6.0,1.0 7.0));

// For Euclidean calculations, the unit of distance is the same as the units
// of your coordinates.
let max_dist = 2.0;
let densified = line_string.densify(&Euclidean, max_dist);
let expected_output = wkt!(LINESTRING(
    0.0 0.0,
    0.0 2.0,
    0.0 4.0,
    0.0 6.0,
    1.0 7.0
));
assert_relative_eq!(densified, expected_output);

For lng/lat geometries, consider using a different metric space like Haversine or Geodesic.

 use geo::wkt;
 use geo::line_measures::{Haversine, Densifiable};
 let line_string = wkt!(LINESTRING(0.0 0.0,0.0 6.0,1.0 7.0));

 // For Haversine, the unit of distance is in meters
 let max_dist = 200_000.0;
 let densified = line_string.densify(&Haversine, max_dist);
 // Haversine interprets coordinate points as lng/lat
 let expected_output = wkt!(LINESTRING(
     0.0 0.0,
     0.0 1.5,
     0.0 3.0,
     0.0 4.5,
     0.0 6.0,
     1.0 7.0
 ));
 assert_relative_eq!(densified, expected_output, epsilon = 1e-14);

Required Associated Types

type Output

Required Methods

fn densify<MetricSpace>( &self, metric_space: &MetricSpace, max_segment_length: F, ) -> Self::Output
where MetricSpace: Distance<F, Point<F>, Point<F>> + InterpolatePoint<F>,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.