Trait geo::algorithm::map_coords::MapCoordsInplace

pub trait MapCoordsInplace<T> {
    fn map_coords_inplace(&mut self, func: impl Fn(&(T, T)) -> (T, T) + Copy)
    where
        T: CoordNum;
}

Map a function over all the coordinates in an object in place

Required methods

fn map_coords_inplace(&mut self, func: impl Fn(&(T, T)) -> (T, T) + Copy) where
T: CoordNum,

Apply a function to all the coordinates in a geometric object, in place

Examples

use geo::algorithm::map_coords::MapCoordsInplace;
use geo::Point;

let mut p = Point::new(10., 20.);
p.map_coords_inplace(|&(x, y)| (x + 1000., y * 2.));

assert_eq!(p, Point::new(1010., 40.));