geo::algorithm::monotone

Function monotone_subdivision

pub fn monotone_subdivision<T: GeoNum, I: IntoIterator<Item = Polygon<T>>>(
    iter: I,
) -> Vec<MonoPoly<T>>

Construct a monotone subdivision (along the X-axis) of an iterator of polygons.

Returns the set of monotone polygons that make up the subdivision. The input polygon(s) must be a valid MultiPolygon (see the validity section in MultiPolygon). In particular, each polygon must be simple, not self-intersect, and contain only finite coordinates; further, the polygons must have distinct interiors, and their boundaries may only intersect at finite points.