geo::algorithm::monotone

Struct MonoPoly

pub struct MonoPoly<T: GeoNum> { /* private fields */ }

Monotone polygon

A monotone polygon is a polygon that can be decomposed into two monotone chains (along the X-axis). This implies any vertical line intersects the polygon at most twice (or not at all). These polygons support point-in-polygon queries in O(log n) time; use the Intersects<Coord> trait to query.

This structure cannot be directly constructed. Use crate::algorithm::monotone_subdivision algorithm to obtain a Vec<MonoPoly>. Consider using MonotonicPolygons instead if you are not interested in the individual monotone polygons.

Implementations

impl<T: GeoNum> MonoPoly<T>

pub fn top(&self) -> &LineString<T>

Get a reference to the mono poly’s top chain.

pub fn bot(&self) -> &LineString<T>

Get a reference to the mono poly’s bottom chain.

pub fn into_ls_pair(self) -> (LineString<T>, LineString<T>)

Convert self to (top, bottom) pair of chains.

pub fn bounding_segment(&self, x: T) -> Option<(Line<T>, Line<T>)>

Get the pair of segments in the chain that intersects the line parallel to the Y-axis at the given x-coordinate. Ties are broken by picking the segment with lower index, i.e. the segment closer to the start of the chains.

pub fn into_polygon(self) -> Polygon<T>

Convert self into a Polygon.