geo::algorithm

Module outlier_detection

Detect outliers in a group of points using LOF

Structs

PreparedDetector

This struct allows multiple detection operations to be run on a point set using varying k_neighbours sizes without having to rebuild the underlying spatial index. Its PreparedDetector::outliers method has the same signature as OutlierDetection::outliers, but retains the underlying spatial index and point set for greater efficiency.

Traits

OutlierDetection

Calculate the Local Outlier Factor of a set of points