geo::algorithm

Module relate

Relate two geometries based on DE-9IM

Structs

GeometryGraph

The computation of the IntersectionMatrix relies on the use of a structure called a “topology graph”. The topology graph contains nodes (CoordNode) and edges (Edge) corresponding to the nodes and line segments of a Geometry. Each node and edge in the graph is labeled with its topological location relative to the source geometry.

IntersectionMatrix

Models a Dimensionally Extended Nine-Intersection Model (DE-9IM) matrix.

PreparedGeometry

A PreparedGeometry can be more efficient than a plain Geometry when performing multiple topological comparisons against the PreparedGeometry.

Traits

Relate

Topologically relate two geometries based on DE-9IM semantics.