geo::algorithm::remove_repeated_points

Trait RemoveRepeatedPoints

pub trait RemoveRepeatedPoints<T>
where
    T: CoordNum + FromPrimitive,
{
    // Required methods
    fn remove_repeated_points(&self) -> Self;
    fn remove_repeated_points_mut(&mut self);
}

Remove repeated points from a MultiPoint and repeated consecutive coordinates from LineString, Polygon, MultiLineString and MultiPolygon.

For GeometryCollection it individually removes the repeated points of each geometry in the collection.

For Point, Line, Rect and Triangle the geometry remains the same.

Required Methods

fn remove_repeated_points(&self) -> Self

Create a new geometry with (consecutive) repeated points removed.

fn remove_repeated_points_mut(&mut self)

Remove (consecutive) repeated points inplace.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.