geo::algorithm::sweep

Trait Cross

pub trait Cross {
    type Scalar: GeoFloat;

    // Required method
    fn line(&self) -> Line<Self::Scalar>;
}

A 1-dimensional finite line used as input to Intersections.

This is implemented by Line, but you can implement it on your own type if you’d like to associate some other data with it.

Cloning

Note that for usage with the Intersections iterator, the type must also impl. Clone. If the custom type is not cheap to clone, use either a reference to the type, a Rc or an Arc. All these are supported via blanket trait implementations.

Required Associated Types

type Scalar: GeoFloat

Scalar used by the line coordinates.

Required Methods

fn line(&self) -> Line<Self::Scalar>

The geometry associated with this type.