geo::algorithm

Trait StitchTriangles

pub trait StitchTriangles<T: GeoFloat>: Stitchable<T> {
    // Required method
    fn stitch_triangulation(
        &self,
    ) -> Result<MultiPolygon<T>, LineStitchingError>;
}

👎Deprecated since 0.32.1: Output is not always valid - use unary_union which is typically faster and produces valid output. Convert your triangles to polygon triangle.to_polygon() first.

Trait to stitch together split up triangles.

Required Methods

fn stitch_triangulation(&self) -> Result<MultiPolygon<T>, LineStitchingError>

👎Deprecated since 0.32.1: Output is not always valid - use unary_union which is typically faster and produces valid output. Convert your triangles to polygon triangle.to_polygon() first.

This stitching only happens along identical edges which are located in two separate geometries. Please read about the required pre conditions of the inputs!

┌─────x        ┌─────┐
│    /│        │     │
│   / │        │     │
│  /  │  ───►  │     │
│ /   │        │     │
│/    │        │     │
x─────┘        └─────┘

Pre Conditions

The triangles in the input must not overlap! This also forbids identical triangles in the input set. If you want to do a union on overlapping triangles then c.f. SpadeBoolops.

Input triangles should be valid polygons. For a definition of validity c.f. https://www.postgis.net/workshops/postgis-intro/validity.html

Examples

use geo::StitchTriangles;
use geo::{Coord, Triangle, polygon};

let tri1 = Triangle::from([
    Coord { x: 0.0, y: 0.0 },
    Coord { x: 1.0, y: 0.0 },
    Coord { x: 0.0, y: 1.0 },
]);
let tri2 = Triangle::from([
    Coord { x: 1.0, y: 1.0 },
    Coord { x: 1.0, y: 0.0 },
    Coord { x: 0.0, y: 1.0 },
]);

let result = vec![tri1, tri2].stitch_triangulation();

assert!(result.is_ok());

let mp = result.unwrap();

assert_eq!(mp.0.len(), 1);

let poly = mp[0].clone();
// 4 coords + 1 duplicate for closed-ness
assert_eq!(poly.exterior().0.len(), 4 + 1);

let expected = polygon![
    Coord { x: 1.0, y: 1.0 },
    Coord { x: 0.0, y: 1.0 },
    Coord { x: 0.0, y: 0.0 },
    Coord { x: 1.0, y: 0.0 },
];

assert_eq!(poly, expected);

Additional Notes

Stitching triangles which result in a polygon with a hole which touches the outline (mentioned here banana polygon) will result in a single polygon without interiors instead of a polygon with a single interior

┌────────x────────┐
│\....../ \....../│
│.\..../   \..../.│
│..\../     \../..│
│...\/       \/...│
│...───────────...│
│../\....^..../\..│
│./..\../.\../..\.│
│/....\/...\/....\│
└─────────────────┘

    │    │    │
    ▼    ▼    ▼

┌────────x────────┐
│       / \       │
│      /   \      │
│     /     \     │
│    /       \    │
│   ───────────   │
│                 │
│                 │
│                 │
└─────────────────┘

If you want to do something more general like a Boolean Operation Union you should use the trait BooleanOps or SpadeBoolops.