geo::algorithm::triangulate_delaunay

Trait TriangulateDelaunay

pub trait TriangulateDelaunay<'a, T>: TriangulationRequirementTrait<'a, T>
where
    T: SpadeTriangulationFloat,
{
    // Provided methods
    fn unconstrained_triangulation(
        &'a self,
    ) -> TriangulationResult<Triangles<T>> { ... }
    fn constrained_outer_triangulation(
        &'a self,
        config: DelaunayTriangulationConfig<T>,
    ) -> TriangulationResult<Triangles<T>> { ... }
    fn constrained_triangulation(
        &'a self,
        config: DelaunayTriangulationConfig<T>,
    ) -> TriangulationResult<Triangles<T>> { ... }
}

Triangulate polygons using a Delaunay Triangulation

This trait contains both constrained and unconstrained triangulation methods. To read more about the differences of these methods also consult this page

Provided Methods

fn unconstrained_triangulation(&'a self) -> TriangulationResult<Triangles<T>>

returns a triangulation thatвҖҷs solely based on the points of the geometric object

The triangulation is guaranteed to be Delaunay

Note that the lines of the triangulation donвҖҷt necessarily follow the lines of the input geometry. If you wish to achieve that take a look at the constrained_triangulation and the constrained_outer_triangulation functions.

use geo::TriangulateDelaunay;
use geo::{Polygon, LineString, Coord};
let u_shape = Polygon::new(
    LineString::new(vec![
        Coord { x: 0.0, y: 0.0 },
        Coord { x: 1.0, y: 0.0 },
        Coord { x: 1.0, y: 1.0 },
        Coord { x: 2.0, y: 1.0 },
        Coord { x: 2.0, y: 0.0 },
        Coord { x: 3.0, y: 0.0 },
        Coord { x: 3.0, y: 3.0 },
        Coord { x: 0.0, y: 3.0 },
    ]),
    vec![],
);
let unconstrained_triangulation = u_shape.unconstrained_triangulation().unwrap();
let num_triangles = unconstrained_triangulation.len();
assert_eq!(num_triangles, 8);

fn constrained_outer_triangulation( &'a self, config: DelaunayTriangulationConfig<T>, ) -> TriangulationResult<Triangles<T>>

returns triangulation thatвҖҷs based on the points of the geometric object and also incorporates the lines of the input geometry

The triangulation is not guaranteed to be Delaunay because of the constraint lines

This outer triangulation also contains triangles that are not included in the input geometry if it wasnвҖҷt convex. HereвҖҷs an example:

в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ\              __/в”Ӯ
в”Ӯ \          __/ / в”Ӯ
в”Ӯ  \      __/   /  в”Ӯ
в”Ӯ   \  __/     /   в”Ӯ
в”Ӯ    \/       /    в”Ӯ
в”Ӯ     в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”җ     в”Ӯ
в”Ӯ    /в”Ӯ\:::::в”Ӯ\    в”Ӯ
в”Ӯ   / в”Ӯ:\::::в”Ӯ \   в”Ӯ
в”Ӯ  /  в”Ӯ::\:::в”Ӯ  \  в”Ӯ
в”Ӯ /   в”Ӯ:::\::в”Ӯ   \ в”Ӯ
в”Ӯ/    в”Ӯ::::\:в”Ӯ    \в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”ҳ______в””в”Җв”Җв”Җв”Җв”Җв”ҳ

use geo::TriangulateDelaunay;
use geo::{Polygon, LineString, Coord};
let u_shape = Polygon::new(
    LineString::new(vec![
        Coord { x: 0.0, y: 0.0 },
        Coord { x: 1.0, y: 0.0 },
        Coord { x: 1.0, y: 1.0 },
        Coord { x: 2.0, y: 1.0 },
        Coord { x: 2.0, y: 0.0 },
        Coord { x: 3.0, y: 0.0 },
        Coord { x: 3.0, y: 3.0 },
        Coord { x: 0.0, y: 3.0 },
    ]),
    vec![],
);
// we use the default [`SpadeTriangulationConfig`] here
let constrained_outer_triangulation =
u_shape.constrained_outer_triangulation(Default::default()).unwrap();
let num_triangles = constrained_outer_triangulation.len();
assert_eq!(num_triangles, 8);

The outer triangulation of the top down U-shape contains extra triangles marked with вҖң:вҖқ. If you want to exclude those, take a look at constrained_triangulation

fn constrained_triangulation( &'a self, config: DelaunayTriangulationConfig<T>, ) -> TriangulationResult<Triangles<T>>

returns triangulation thatвҖҷs based on the points of the geometric object and also incorporates the lines of the input geometry

The triangulation is not guaranteed to be Delaunay because of the constraint lines

This triangulation only contains triangles that are included in the input geometry. HereвҖҷs an example:

в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ\              __/в”Ӯ
в”Ӯ \          __/ / в”Ӯ
в”Ӯ  \      __/   /  в”Ӯ
в”Ӯ   \  __/     /   в”Ӯ
в”Ӯ    \/       /    в”Ӯ
в”Ӯ     в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”җ     в”Ӯ
в”Ӯ    /в”Ӯ      в”Ӯ\    в”Ӯ
в”Ӯ   / в”Ӯ      в”Ӯ \   в”Ӯ
в”Ӯ  /  в”Ӯ      в”Ӯ  \  в”Ӯ
в”Ӯ /   в”Ӯ      в”Ӯ   \ в”Ӯ
в”Ӯ/    в”Ӯ      в”Ӯ    \в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”ҳ      в””в”Җв”Җв”Җв”Җв”Җв”ҳ

use geo::TriangulateDelaunay;
use geo::{Polygon, LineString, Coord};
let u_shape = Polygon::new(
    LineString::new(vec![
        Coord { x: 0.0, y: 0.0 },
        Coord { x: 1.0, y: 0.0 },
        Coord { x: 1.0, y: 1.0 },
        Coord { x: 2.0, y: 1.0 },
        Coord { x: 2.0, y: 0.0 },
        Coord { x: 3.0, y: 0.0 },
        Coord { x: 3.0, y: 3.0 },
        Coord { x: 0.0, y: 3.0 },
    ]),
    vec![],
);
// we use the default [`DelaunayTriangulationConfig`] here
let constrained_triangulation = u_shape.constrained_triangulation(Default::default()).unwrap();
let num_triangles = constrained_triangulation.len();
assert_eq!(num_triangles, 6);

Compared to the constrained_outer_triangulation it only includes the triangles inside of the input geometry

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.