geo::algorithm

Module validation

Provides a way to check the validity of geometries, based on the OGC Simple Feature Access - Part 1: Common Architecture standard.

Structs

CoordIndex

The index of the coordinate in the geometry

GeometryIndex

The position of the problem in a multi-geometry, starting at 0.

Enums

InvalidGeometry

A Geometry is valid if its inner variant is valid. e.g. Geometry::Polygon(polygon) is valid if and only if polygon is valid.

InvalidGeometryCollection

A GeometryCollection is valid if all its elements are valid.

InvalidLine

InvalidLineString

InvalidMultiLineString

A MultiLineString is valid if each LineString in it is valid.

InvalidMultiPoint

A MultiPoint is valid if each Point in it is valid.

InvalidMultiPolygon

A MultiPolygon is valid if:

InvalidPoint

InvalidPolygon

A Polygon must follow these rules to be valid:

InvalidRect

InvalidTriangle

RingRole

The role of a ring in a Polygon.

Traits

Validation

A trait to check if a geometry is valid and report the reason(s) of invalidity.