geo::algorithm

Module winding_order

Calculate and work with the winding order of Linestrings.

Structs

Points

Iterates through a list of Points

Enums

WindingOrder

How a linestring is wound, clockwise or counter-clockwise

Traits

Winding

Determine and operate on how a LineString is wound. This functionality, and our implementation is based on CGAL’s Polygon_2::orientation.

Functions

triangle_winding_order

special cased algorithm for finding the winding of a triangle