Struct geo::MultiPolygon

pub struct MultiPolygon<T>(pub Vec<Polygon<T>, Global>)
 where
    T: CoordNum;

A collection of Polygons. Can be created from a Vec of Polygons, or from an Iterator which yields Polygons. Iterating over this object yields the component Polygons.

Semantics

The interior and the boundary are the union of the interior and the boundary of the constituent polygons.

Validity

The interiors of no two constituent polygons may intersect.

The boundaries of two (distinct) constituent polygons may only intersect at finitely many points.

Refer to section 6.1.14 of the OGC-SFA for a formal definition of validity. Note that the validity is not enforced, but expected by the operations and predicates that operate on it.

Tuple Fields
0: Vec<Polygon<T>, Global>
Implementations

impl<T> MultiPolygon<T> where
T: CoordNum,

pub fn iter(&self) -> impl Iterator<Item = &Polygon<T>>

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut Polygon<T>>