geo

Type Alias Coordinate

pub type Coordinate<T = f64> = Coord<T>;

👎Deprecated: Renamed to geo_types::Coord (or geo::Coord)

Aliased Type

pub struct Coordinate<T = f64> {
    pub x: T,
    pub y: T,
}

Fields
x: T

Typically, x is the horizontal position, or longitude for geographic coordinates, but its interpretation can vary across coordinate systems.

y: T

Typically, y is the vertical position, or latitude for geographic coordinates, but its interpretation can vary across coordinate systems.