Crate geo

Re-exports

pub use algorithm::*;

Modules

algorithm
This module includes all the functions of geometric calculations

prelude
A prelude which re-exports the traits for manipulating objects in this crate. Typically imported with use geo::prelude::*.

Structs

Bbox

Coordinate

ExtremePoint

Extremes

GeometryCollection
A collection of Geometrys.

Line

LineString
A LineString, which is an ordered collection of Points.

MultiLineString
A collection of LineStrings.

MultiPoint
A collection of Points.

MultiPolygon
A collection of Polygons.

Point
A single Point in 2D space.

Polygon
A 2D polygon area.

Enums

Closest
The result of trying to find the closest spot on an object to a point.

Geometry
An enum representing any possible geomtry type.

Statics

COORD_PRECISION

Traits

ToGeo