Module geo::algorithm

This module includes all the functions of geometric calculations

Modules

area
Returns the area of the surface of a geometry.

boundingbox
Returns the Bbox of a geometry.

centroid
Calculation of the centroid.

closest_point
Determine the closest point between two objects.

contains
Checks if the geometry A is completely inside the B geometry.

convexhull
Calculates the convex hull of a geometry.

distance
Returns the Euclidean distance between two geometries.

extremes
Returns the extreme indices of a Polygon, MultiPolygon, or MultiPoint.

haversine_destination
Returns a new Point using distance and bearing.

haversine_distance
Returns the Haversine distance between two geometries.

intersects
Checks if the geometry A intersects the geometry B.

length
Returns the length of a line.

map_coords
Apply a function to all coordinates

orient
Orients a Polygon's exterior and interior rings.

rotate
Rotates a geometry around either its centroid or a point by an angle, given in degrees.

simplify
Simplifies geometries using the Ramer-Douglas-Peucker algorithm.

simplifyvw
Simplifies geometries using the Visvalingam-Whyatt algorithm. Includes a topology-preserving variant.

translate
Translates a geometry along the given offsets.