Module geo::algorithm

This module includes all the functions of geometric calculations

Modules

area
Returns the area of the surface of geometry

boundingbox
Returns the Bbox of a geometry.

centroid
Calculation of the centroid.

contains
Checks if the geometry A is completely inside the B geometry.

distance
Returns the distance between two geometries.

intersects
hecks if the geometry A intersects the geometry B.

length
Returns the length of a line.

simplify
Simplifies a LineString using the Ramer-Douglas-Peucker algorithm.

simplifyvw
Simplifies a LineString using the Visvalingam-Whyatt algorithm.