geo::algorithm::haversine_intermediate

Trait HaversineIntermediate

pub trait HaversineIntermediate<T: CoordFloat> {
    // Required methods
    fn haversine_intermediate(&self, other: &Point<T>, ratio: T) -> Point<T>;
    fn haversine_intermediate_fill(
        &self,
        other: &Point<T>,
        max_dist: T,
        include_ends: bool,
    ) -> Vec<Point<T>>;
}

👎Deprecated since 0.29.0: Please use the InterpolatePoint trait instead

Returns a new Point along a great circle route between two existing points

Required Methods

fn haversine_intermediate(&self, other: &Point<T>, ratio: T) -> Point<T>

👎Deprecated since 0.29.0: Please use Haversine.point_at_ratio_between from the InterpolatePoint trait instead

Returns a new Point along a great circle route between self and other.

other - The other point to interpolate towards.

ratio - How far along the route should the new point be, with 0.0 being at self and 1.0 being at other.

Examples

use geo::HaversineIntermediate;
use geo::Point;

let p1 = Point::new(10.0, 20.0);
let p2 = Point::new(125.0, 25.0);

let i20 = p1.haversine_intermediate(&p2, 0.2);
assert_relative_eq!(i20, Point::new(29.8, 29.9), epsilon = 0.2);

let i80 = p1.haversine_intermediate(&p2, 0.8);
assert_relative_eq!(i80, Point::new(103.5, 33.5), epsilon = 0.2);

fn haversine_intermediate_fill( &self, other: &Point<T>, max_dist: T, include_ends: bool, ) -> Vec<Point<T>>

👎Deprecated since 0.29.0: Please use Haversine.points_along_line from the InterpolatePoint trait instead

Interpolates Points along a great circle route between self and other.

As many points as necessary will be added such that the distance between points never exceeds max_dist.

include_ends: Should the start and end points be included in the output?