geo::algorithm::line_measures::metric_spaces

Struct Euclidean

pub struct Euclidean;

Operations on the Euclidean plane measure distance with the pythagorean formula - what you’d measure with a ruler.

If you have lon/lat points, use the Haversine, Geodesic, or other metric spaces - Euclidean methods will give nonsense results.

If you wish to use Euclidean operations with lon/lat, the coordinates must first be transformed using the Transform::transform / Transform::transform_crs_to_crs methods or their immutable variants. Use of these requires the proj feature