geo::algorithm

Module linestring_segment

Split a LineString into n segments

Traits

LineStringSegmentize

Segments a LineString into segment_count equal length LineStrings as a MultiLineString using Euclidean distance calculations. See LineStringSegmentizeHaversine if you are dealing with geographic coordinates (lat/lon).

LineStringSegmentizeHaversine

Segments a LineString into segment_count equal length LineStrings as a MultiLineString using Haversine distance calculations. Use this over LineStringSegmentize when using data from a geographic coordinate system.