geo::algorithm::relate

Struct IntersectionMatrix

pub struct IntersectionMatrix(/* private fields */);

Models a Dimensionally Extended Nine-Intersection Model (DE-9IM) matrix.

DE-9IM matrix values (such as “212FF1FF2”) specify the topological relationship between two Geometeries.

DE-9IM matrices are 3x3 matrices that represent the topological locations that occur in a geometry (Interior, Boundary, Exterior).

The indices are provided by the enum cases CoordPos::Inside, CoordPos::OnBoundary, CoordPos::Outside.

The matrix entries represent the Dimensions of each intersection.

For a description of the DE-9IM and the spatial predicates derived from it, see the following references:

OGC 99-049 OpenGIS Simple Features Specification for SQL, Section 2.1.13

OGC 06-103r4 OpenGIS Implementation Standard for Geographic information - Simple feature access - Part 1: Common architecture, Section 6.1.15 (which provides some further details on certain predicate specifications).

Wikipedia article on DE-9IM

This implementation is heavily based on that from the JTS project.

Implementations

impl IntersectionMatrix

pub fn empty() -> Self

pub fn is_disjoint(&self) -> bool

Tests if this matrix matches [FF*FF****].

returns true if the two geometries related by this matrix are disjoint

pub fn is_intersects(&self) -> bool

Tests if is_disjoint returns false.

returns true if the two geometries related by this matrix intersect.

pub fn is_within(&self) -> bool

Tests whether this matrix matches [T*F**F***].

returns true if the first geometry is within the second.

pub fn is_contains(&self) -> bool

Tests whether this matrix matches [T*****FF*].

returns true if the first geometry contains the second.