geo::algorithm

Module rhumb

Rhumb-line-related algorithms and utils This module provides rhumb-line (a.k.a. loxodrome) geometry operations. The distance, destination, and bearing implementations are adapted in part from their equivalents in Turf.js, which in turn are adapted from the Movable Type spherical geodesy tools. Turf.js is copyright its authors and the geodesy tools are copyright Chris Veness; both are available under an MIT license.

Traits

RhumbBearingDeprecated

Returns the bearing to another Point in degrees.

RhumbDestinationDeprecated

Returns the destination Point having travelled the given distance along a [rhumb line] from the origin geometry with the given bearing

RhumbDistanceDeprecated

Determine the distance between two geometries along a rhumb line.

RhumbIntermediateDeprecated

Returns a new Point along a rhumb line between two existing points

RhumbLengthDeprecated

Determine the length of a geometry assuming each segment is a rhumb line.