geo::algorithm::sweep

Trait Cross

pub trait Cross: Sized + Debug {
    type Scalar: GeoFloat;

    // Required method
    fn line(&self) -> LineOrPoint<Self::Scalar>;
}

Interface for types that can be processed to detect crossings.

This type is implemented by LineOrPoint, but users may also implement this on custom types to store extra information. Any type that represents an ordered line-segment may implement this.

Cloning

Note that for usage with the planar sweep iterators, the type must also impl. Clone. If the custom type is not cheap to clone, use either a reference to the type, a Rc or an Arc. All these are supported via blanket trait implementations.

Required Associated Types

type Scalar: GeoFloat

Scalar used the coordinates.

Required Methods

fn line(&self) -> LineOrPoint<Self::Scalar>

The geometry associated with this type. Use a Line with the start and end coordinates to represent a point.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.