geo::algorithm::translate

Trait Translate

pub trait Translate<T: CoordNum> {
    // Required methods
    fn translate(&self, x_offset: T, y_offset: T) -> Self;
    fn translate_mut(&mut self, x_offset: T, y_offset: T);
}

Required Methods

fn translate(&self, x_offset: T, y_offset: T) -> Self

Translate a Geometry along its axes by the given offsets

Performance

If you will be performing multiple transformations, like Scale, Skew, Translate, or Rotate, it is more efficient to compose the transformations and apply them as a single operation using the AffineOps trait.

Examples

use geo::Translate;
use geo::line_string;

let ls = line_string![
    (x: 0.0, y: 0.0),
    (x: 5.0, y: 5.0),
    (x: 10.0, y: 10.0),
];

let translated = ls.translate(1.5, 3.5);

assert_eq!(translated, line_string![
    (x: 1.5, y: 3.5),
    (x: 6.5, y: 8.5),
    (x: 11.5, y: 13.5),
]);

fn translate_mut(&mut self, x_offset: T, y_offset: T)

Translate a Geometry along its axes, but in place.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.