geo::geometry

Struct Coord

pub struct Coord<T = f64>
where
    T: CoordNum,
{
    pub x: T,
    pub y: T,
}

A lightweight struct used to store coordinates on the 2-dimensional Cartesian plane.

Unlike Point (which in the future may contain additional information such as an envelope, a precision model, and spatial reference system information), a Coord only contains ordinate values and accessor methods.

This type implements the vector space operations: Add, Sub, Neg, Zero, Mul<T>, and Div<T> traits.

Semantics

This type does not represent any geospatial primitive, but is used in their definitions. The only requirement is that the coordinates it contains are valid numbers (for eg. not f64::NAN).

Fields
x: T

Typically, x is the horizontal position, or longitude for geographic coordinates, but its interpretation can vary across coordinate systems.

y: T

Typically, y is the vertical position, or latitude for geographic coordinates, but its interpretation can vary across coordinate systems.

Implementations

impl<T> Coord<T>
where T: CoordNum,

pub fn x_y(&self) -> (T, T)

Returns a tuple that contains the x/horizontal & y/vertical component of the coordinate.

Examples

use geo_types::coord;

let c = coord! {
    x: 40.02f64,
    y: 116.34,
};
let (x, y) = c.x_y();

assert_eq!(y, 116.34);
assert_eq!(x, 40.02f64);

impl<T> Coord<T>
where T: CoordNum,

Create a coordinate at the origin.

Examples

use geo_types::Coord;
use num_traits::Zero;

let p: Coord = Zero::zero();

assert_eq!(p.x, 0.);
assert_eq!(p.y, 0.);

pub fn zero() -> Coord<T>