Struct geo::geometry::Line

pub struct Line<T = f64>where
    T: CoordNum,{
    pub start: Coord<T>,
    pub end: Coord<T>,
}

A line segment made up of exactly two Coords.

Semantics

The interior and boundary are defined as with a LineString with the two end points.

Fields
start: Coord<T>end: Coord<T>
Implementations

impl<T> Line<T>where
T: CoordNum,

pub fn new<C>(start: C, end: C) -> Line<T>where
C: Into<Coord<T>>,

Creates a new line segment.

Examples

use geo_types::{coord, Line};

let line = Line::new(coord! { x: 0., y: 0. }, coord! { x: 1., y: 2. });

assert_eq!(line.start, coord! { x: 0., y: 0. });
assert_eq!(line.end, coord! { x: 1., y: 2. });

pub fn delta(&self) -> Coord<T>

Calculate the difference in coordinates (Δx, Δy).

pub fn dx(&self) -> T

Calculate the difference in ‘x’ components (Δx).

Equivalent to:

line.end.x - line.start.x

pub fn dy(&self) -> T

Calculate the difference in ‘y’ components (Δy).

Equivalent to:

line.end.y - line.start.y

pub fn slope(&self) -> T

Calculate the slope (Δy/Δx).

Equivalent to:

line.dy() / line.dx()

Note that:

Line::new(a, b).slope() == Line::new(b, a).slope()

pub fn determinant(&self) -> T

Calculate the determinant of the line.

Equivalent to:

line.start.x * line.end.y - line.start.y * line.end.x

Note that:

Line::new(a, b).determinant() == -Line::new(b, a).determinant()

pub fn start_point(&self) -> Point<T>

pub fn end_point(&self) -> Point<T>

pub fn points(&self) -> (Point<T>, Point<T>)