Struct geo::geometry::MultiLineString

pub struct MultiLineString<T = f64>(pub Vec<LineString<T>, Global>)
where
    T: CoordNum;

A collection of LineStrings. Can be created from a Vec of LineStrings or from an Iterator which yields LineStrings. Iterating over this object yields the component LineStrings.

Semantics

The boundary of a MultiLineString is obtained by applying the “mod 2” union rule: A Point is in the boundary of a MultiLineString if it is in the boundaries of an odd number of elements of the MultiLineString.

The interior of a MultiLineString is the union of the interior, and boundary of the constituent LineStrings, except for the boundary as defined above. In other words, it is the set difference of the boundary from the union of the interior and boundary of the constituents.

A MultiLineString is simple if and only if all of its elements are simple and the only intersections between any two elements occur at Points that are on the boundaries of both elements. A MultiLineString is closed if all of its elements are closed. The boundary of a closed MultiLineString is always empty.

Tuple Fields
0: Vec<LineString<T>, Global>
Implementations

impl<T> MultiLineString<T>where
T: CoordNum,

pub fn new(value: Vec<LineString<T>, Global>) -> MultiLineString<T>

Instantiate Self from the raw content value

pub fn is_closed(&self) -> bool

True if the MultiLineString is empty or if all of its LineStrings are closed - see LineString::is_closed.

Examples

use geo_types::{MultiLineString, LineString, line_string};

let open_line_string: LineString<f32> = line_string![(x: 0., y: 0.), (x: 5., y: 0.)];
assert!(!MultiLineString::new(vec![open_line_string.clone()]).is_closed());

let closed_line_string: LineString<f32> = line_string![(x: 0., y: 0.), (x: 5., y: 0.), (x: 0., y: 0.)];
assert!(MultiLineString::new(vec![closed_line_string.clone()]).is_closed());

// MultiLineString is not closed if *any* of it's LineStrings are not closed
assert!(!MultiLineString::new(vec![open_line_string, closed_line_string]).is_closed());

// An empty MultiLineString is closed
assert!(MultiLineString::<f32>::new(vec![]).is_closed());

impl<T> MultiLineString<T>where
T: CoordNum,

pub fn iter(&self) -> impl Iterator<Item = &LineString<T>>

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut LineString<T>>