geo::geometry

Struct Triangle

pub struct Triangle<T = f64>(pub Coord<T>, pub Coord<T>, pub Coord<T>)
where
    T: CoordNum;

A bounded 2D area whose three vertices are defined by Coords. The semantics and validity are that of the equivalent Polygon; in addition, the three vertices must not be collinear and they must be distinct.

Notes

Irrespective of input order the resulting geometry has ccw order and its vertices are yielded in ccw order by iterators

Tuple Fields
0: Coord<T>1: Coord<T>2: Coord<T>
Implementations

impl<T> Triangle<T>
where T: CoordNum,

pub fn new(v1: Coord<T>, v2: Coord<T>, v3: Coord<T>) -> Triangle<T>

Instantiate Self from the raw content value

pub fn to_array(&self) -> [Coord<T>; 3]

pub fn to_lines(&self) -> [Line<T>; 3]

pub fn to_polygon(self) -> Polygon<T>

Create a Polygon from the Triangle.

Examples

use geo_types::{coord, Triangle, polygon};

// Input is CW
let triangle = Triangle::new(
    coord! { x: 0., y: 0. },
    coord! { x: 10., y: 20. },
    coord! { x: 20., y: -10. },
);

// Output is CCW
assert_eq!(
    triangle.to_polygon(),
    polygon![
        (x: 20., y: -10.),
        (x: 10., y: 20.),
        (x: 0., y: 0.),
        (x: 20., y: -10.),
    ],
);