Struct geo::Line

pub struct Line<T> where
    T: CoordNum,  {
    pub start: Coordinate<T>,
    pub end: Coordinate<T>,
}

A line segment made up of exactly two Coordinates.

Semantics

The interior and boundary are defined as with a LineString with the two end points.

Fields
start: Coordinate<T>end: Coordinate<T>
Implementations

impl<T> Line<T> where
T: CoordNum,

pub fn new<C>(start: C, end: C) -> Line<T> where
C: Into<Coordinate<T>>,

Creates a new line segment.

Examples

use geo_types::{Coordinate, Line};

let line = Line::new(Coordinate { x: 0., y: 0. }, Coordinate { x: 1., y: 2. });

assert_eq!(line.start, Coordinate { x: 0., y: 0. });
assert_eq!(line.end, Coordinate { x: 1., y: 2. });

pub fn delta(&self) -> Coordinate<T>

Calculate the difference in coordinates (Δx, Δy).

pub fn dx(&self) -> T

Calculate the difference in ‘x’ components (Δx).

Equivalent to:

line.end.x - line.start.x

pub fn dy(&self) -> T

Calculate the difference in ‘y’ components (Δy).

Equivalent to:

line.end.y - line.start.y

pub fn slope(&self) -> T

Calculate the slope (Δy/Δx).

Equivalent to:

line.dy() / line.dx()

Note that:

Line::new(a, b).slope() == Line::new(b, a).slope()

pub fn determinant(&self) -> T

Calculate the determinant of the line.

Equivalent to:

line.start.x * line.end.y - line.start.y * line.end.x

Note that:

Line::new(a, b).determinant() == -Line::new(b, a).determinant()

pub fn start_point(&self) -> Point<T>

pub fn end_point(&self) -> Point<T>

pub fn points(&self) -> (Point<T>, Point<T>)