geo

Trait GeoNum

pub trait GeoNum: CoordNum {
    type Ker: Kernel<Self>;

    // Required method
    fn total_cmp(&self, other: &Self) -> Ordering;
}

A trait for methods which work for both integers and floating point

Required Associated Types

type Ker: Kernel<Self>

Required Methods

fn total_cmp(&self, other: &Self) -> Ordering

Return the ordering between self and other.

For integers, this should behave just like Ord.

For floating point numbers, unlike the standard partial comparison between floating point numbers, this comparison always produces an ordering.

See f64::total_cmp for details.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.